+++ HISTORY.rst
@@ -1,5 +1,13 @@
 Changelog
 ---------
+
+- Adds a helper for getting the title translations.
+  [msom]
+
+- Uses the default locale of the site to auto-generate the ids of elections
+  and votes.
+  [msom]
+
 1.10.3 (2018-01-15)
 ~~~~~~~~~~~~~~~~~~~